fix(objetivos): Change resultado attribute to Float

+++ talent360/evaluaciones/models/objetivo.py
@@ -87,7 +87,7 @@ class Objetivo(models.Model):
         default=fields.Datetime.today(),
         help="Fecha en la que se debe cumplir el objetivo",
     )
-    resultado = fields.Integer(store=True)
+    resultado = fields.Float(store=True)
     porcentaje = fields.Float(store=True)
     estado = fields.Selection(
         [
